fix: Use model name instead of local path for AutoProcessor.from_pretrained
Fixes AttributeError when do_formula_enrichment=True. When a local Path
is passed to AutoProcessor.from_pretrained(), transformers loads the
config as a dict but then tries to access .model_type as an attribute.

Using the model name ('docling-project/CodeFormulaV2') allows
transformers to properly load the config as an object, while still
using the cached model automatically.

Fixes #2681

Signed-off-by: bryan <b-g-d@users.noreply.github.com>